+++ src/view/resourceView.ts
@@ -3,76 +3,50 @@ import * as vscode from 'vscode';
 import _ from 'lodash';
 
 import { Node, NodeState, ResourceGres } from '../models/';
-import { ListItem, NodeItem } from './components';
+import { GresItem, NodeItem } from './components';
 import { configService, GresSortKeys, GresGroupKeys, resourceService } from '../services';
 import { resignFn } from '../utils/utils';
 
-function gresIcon(gres: ResourceGres): vscode.ThemeIcon {
-    if (gres.totalNum === 0) { return new vscode.ThemeIcon('close', new vscode.ThemeColor('gitDecoration.deletedResourceForeground')); }
-    if (gres.usedNum === 0) { return new vscode.ThemeIcon('circle-filled', new vscode.ThemeColor('gitDecoration.addedResourceForeground')); }
-    if (gres.usedNum < gres.totalNum) { return new vscode.ThemeIcon('circle-filled', new vscode.ThemeColor('gitDecoration.modifiedResourceForeground')); }
-    return new vscode.ThemeIcon('circle-filled', new vscode.ThemeColor('gitDecoration.deletedResourceForeground'));
-}
 
-const nodeSortFn = new Map([
+const nodeSortFnMap = new Map([
     [GresSortKeys.NAME, (a: Node, b: Node) => a.nodeid.localeCompare(b.nodeid, undefined, { sensitivity: 'base' })],
     [GresSortKeys.AVAIL, (a: Node, b: Node) => a.gres === null || b.gres === null ? 0 : (!a.isAvailableState ? -1 : (!b.isAvailableState ? 1 : (a.gres.totalNum - a.gres.usedNum) - (b.gres.totalNum - b.gres.usedNum)))],
 ]);
 
-const listSortFn = new Map([
+const gresSortFnMap = new Map([
     [GresSortKeys.NAME, (a: ResourceGres | null, b: ResourceGres | null) => a === null ? -1 : (b === null ? 1 : a.toIdString().localeCompare(b.toIdString(), undefined, { sensitivity: 'base' }))],
     [GresSortKeys.AVAIL, (a: ResourceGres | null, b: ResourceGres | null) => a === null ? -1 : (b === null ? 1 : (a.totalNum - a.usedNum) - (b.totalNum - b.usedNum))],
 ]);
 
-function getNodeGroupedByGRES(): ListItem[] {
-    return Object.values(_.groupBy(resourceService.getNode(), node => node.gres?.toIdString() ?? 'null'))
-        .map(nodes => {
-            nodes = nodes.sort(resignFn(nodeSortFn.get(configService.gresSortKey)!, configService.gresSortAscending));
-            if (nodes[0].gres === null) {
-                return [null, new ListItem('No GRES', nodes.map(v => new NodeItem(v)), vscode.l10n.t('${length} nodes'), undefined, 'gresList', false)];
-            }
-            const availNode = nodes.filter(v => v.isAvailableState);
-            const rgres = availNode.length === 0 ? ResourceGres.empty(nodes[0]!.gres.toIdString()) : ResourceGres.fromArray(availNode.map(v => v.gres!));
+function getNodeGroupedByGres(): GresItem[] {
+    const gresNode = resourceService.getNodesByGres();
 
-            const item = new ListItem(rgres.toString(), nodes.map(v => new NodeItem(v)), vscode.l10n.t('${length} nodes'), gresIcon(rgres), 'gresList', false);
-            item.tooltip = new vscode.MarkdownString(`
-| nodeid | state | GRES (idle/total) | memory (i/t) | CPUs (i/t) |
-|:--:|:--:|:--:|:--:|:--:|
-${nodes.slice(0, 10).map(node => `| ${node.nodeid} | ${node.state} | ${node.gres ?? 'No GRES'} | ${Math.round((node.memory - node.allocMemory) * 1000) / 1000}GB / ${node.memory}GB | ${node.idleCpu} / ${node.cpu} |`).join('\n')}
-${nodes.length > 10 ? '|...|...|...|...|...|' : ''}
-            `);
-            return [rgres, item];
-        })
-        .sort((a, b) => resignFn(listSortFn.get(configService.gresSortKey)!, configService.gresSortAscending)(a[0] as ResourceGres | null, b[0] as ResourceGres | null))
-        .map(item => item[1]) as ListItem[];
+    const gresSortFn = resignFn(gresSortFnMap.get(configService.gresSortKey)!, configService.gresSortAscending);
+    const sortedMap = [...gresNode].sort((a, b) => gresSortFn(a[1].gres, b[1].gres));
+
+    return sortedMap.map(([gresId, group]) => {
+        return new GresItem(group.gres, group.nodes);
+    });
 }
 
-function getNodeGroupedByPartition(): ListItem[] {
-    return Object.values(_.groupBy(resourceService.getNode(), node => node.partition))
-        .map(nodes => {
-            nodes = nodes.sort(resignFn(nodeSortFn.get(configService.gresSortKey)!, configService.gresSortAscending));
-            const availNode = nodes.filter(v => v.isAvailableState && v.gres !== null);
-            const rgres = availNode.length === 0 ? ResourceGres.empty('') : ResourceGres.fromArray(availNode.map(v => v.gres!));
+function getNodeGroupedByPartition(): GresItem[] {
+    const partitionNode = resourceService.getNodeByPartition();
 
-            const item = new ListItem(`${nodes[0].partition} (${rgres.totalNum - rgres.usedNum}/${rgres.totalNum})`, nodes.map(v => new NodeItem(v)), vscode.l10n.t('${length} nodes'), gresIcon(rgres), 'gresList', false);
-            item.tooltip = new vscode.MarkdownString(`
-| nodeid | state | GRES (idle/total) | memory (i/t) | CPUs (i/t) |
-|:--:|:--:|:--:|:--:|:--:|
-${nodes.slice(0, 10).map(node => `| ${node.nodeid} | ${node.state} | ${node.gres ?? 'No GRES'} | ${Math.round((node.memory - node.allocMemory) * 1000) / 1000}GB / ${node.memory}GB | ${node.idleCpu} / ${node.cpu} |`).join('\n')}
-${nodes.length > 10 ? '|...|...|...|...|...|' : ''}
-            `);
-            return [rgres, item];
-        })
-        .sort((a, b) => resignFn(listSortFn.get(configService.gresSortKey)!, configService.gresSortAscending)(a[0] as ResourceGres | null, b[0] as ResourceGres | null))
-        .map(item => item[1]) as ListItem[];
+    const gresSortFn = resignFn(gresSortFnMap.get(configService.gresSortKey)!, configService.gresSortAscending);
+
+    const sortedMap = [...partitionNode].sort((a, b) => gresSortFn(a[1].gres, b[1].gres));
+
+    return sortedMap.map(([partition, group]) => {
+        return new GresItem(group.gres, group.nodes, `${partition} (${group.gres!.totalNum - group.gres!.usedNum}/${group.gres!.totalNum})`);
+    });
 }
 
 const getGroupedNodeFn = {
-    [GresGroupKeys.GRES]: getNodeGroupedByGRES,
+    [GresGroupKeys.GRES]: getNodeGroupedByGres,
     [GresGroupKeys.PARTITION]: getNodeGroupedByPartition,
 };
 
-export class ResourceViewDataProvider implements vscode.TreeDataProvider<NodeItem | ListItem> {
+export class ResourceViewDataProvider implements vscode.TreeDataProvider<NodeItem | GresItem> {
     private static _instance: ResourceViewDataProvider | null = null;
     private constructor() { }
 
@@ -90,23 +64,24 @@ export class ResourceViewDataProvider implements vscode.TreeDataProvider<NodeIte
         this._onDidChangeTreeData.fire();
     }
 
-    getTreeItem(element: NodeItem | ListItem): vscode.TreeItem {
+    getTreeItem(element: NodeItem | GresItem): vscode.TreeItem {
         return element;
     }
 
-    getChildren(element?: NodeItem | ListItem): Thenable<ListItem[] | NodeItem[]> {
+    getChildren(element?: NodeItem | GresItem): Thenable<GresItem[] | NodeItem[]> {
         if (!element) {
             return Promise.resolve(getGroupedNodeFn[configService.gresGroupKey]());
         }
-        if (element instanceof ListItem) {
-            return Promise.resolve(element.children as NodeItem[]);
+        if (element instanceof GresItem) {
+            const nodeSortFn = resignFn(nodeSortFnMap.get(configService.gresSortKey)!, configService.gresSortAscending);
+            return Promise.resolve(element.nodes.sort(nodeSortFn).map(v => new NodeItem(v)));
         }
         return Promise.resolve([]);
     }
 }
 
 export const resourceViewDataProvider = ResourceViewDataProvider.getInstance();
-export let resourceTreeView: vscode.TreeView<ListItem | NodeItem>;
+export let resourceTreeView: vscode.TreeView<GresItem | NodeItem>;
 
 export function initResourceView(context: vscode.ExtensionContext) {
     resourceTreeView = vscode.window.createTreeView('slurm--_resource_view', { treeDataProvider: resourceViewDataProvider });